requirements.txt
asn1crypto==1.5.1
bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
fuse-python==1.0.7
gpg==1.18.0
html5lib==1.1
httplib2==0.20.4
idna==3.6
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Markdown==3.5.2
markdown-it-py==3.0.0
mdurl==0.1.2
mechanize==0.4.9
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pylibacl==0.7.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-apt==2.7.7+ubuntu5
python-debian==0.1.49+ubuntu2
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
systemd-python==235
tornado==6.4
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
vboxapi==1
wadllib==1.3.6
webencodings==0.5.1
wheel==0.42.0
xdg==5
xkit==0.0.0
